requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
cross-web==0.4.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
fastapi==0.135.1
graphql-core==3.2.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
motor==3.7.1
packaging==26.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
strawberry-graphql==0.307.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
